Remove euclid & num_traits
