+++ gdn/ownwar/src/util/mod.rs
@@ -2,12 +2,10 @@
 
 mod bitarray;
 mod iter_n_dim;
-mod vector;
 mod voxel_raycast;
 mod voxel_sphere_iter;
 
 pub use bitarray::BitArray;
 pub use iter_n_dim::*;
-pub use vector::convert_vec;
 pub use voxel_raycast::VoxelRaycast;
 pub use voxel_sphere_iter::VoxelSphereIterator;
